requirements.txt
streamlit
pdfminer.six
python-docx
openai  # for LLM generation (optional for later)
